Basic packet handling
Finished up the packet handling on the Coordinator layer to handle PacketType of INITALIZATION from the server. The server sends it's preferred listening port and the coordinator saves this and puts it into the config file. After this the Coordinator will send a packet of the type ACK back to the sender to close the connection.

If the server gets anything other than a packet type of ACK during this process, it will wait a second, retry. It does this 3 times and then will shut down if unsuccessful.