PacketHandling Class
Created an Abstract Superclass for the packet handling.
Since the packets are all going to have their payloads torn into key/value pairs the handling SuperClass will handle that aspect of receiving packets.

Then the actual processing of the key/value pairs will be overwritten and handling on a case by case basis.

I also added the Termination line and proper handling for the lack of a termination line.

Also added handling for an empty payload as well