Server Initialization ACK handling
Added better handling for the server's initialization and the response packet.

Will throw an exception if the packet is not an ACK packet or does not have the proper termination of the packet.

Will retry connection if there is anything other than an ACK packet sent back. Has a retry limit of 3 times and then will terminate if that limit is reached