Finished up packet DTO reform and started DEMO
- Added handling for packetType MESSAGE
- Fixed issue with packet recreation upon receiving with Gson. Since Abstract classes could not be created with Gson.fromJson(), I needed to use the RuntimeTypeAdapterFactory in order to register a 'generic' packet type with a different subtype - this threw an issue with registering same type packets (e.g. genericPacket) with differing subtypes
----(for example,
---------register(genericPacket, INIT);
---------register(genericPacket, MESSAGE);
would throw an error since they are of the same type. So I created a dynamic work around where the PacketType is pulled preemptively from the json and then passed to the RuntimeTypeAdapterFactory. This works since any instance of packetHandler is only ran once for that packet.

Other commenting fixes