Fixed VSCode and batch file issue
Batch file would compile project independently of Maven. Project would compile due to the batch file compiling using Javac. VSCode was also throwing a package error and import errors. When I fixed the VS Code error the compilation would not work. Due to the batch file error the package issue would result it it not being able to be found